Implementation
The data has been split into two groups:
1. Training set
2. Test set
The training set is used to build ML models. For the training set, the outcome (also known as the “ground truth”) for each passenger is given. The model is based on “features” like passengers’ gender and class. I have use feature engineering to create new features.

The test set is being used to see how well our model performs on unseen data. For the test set, the ground truth for each passenger will not be present. It is our job to predict these outcomes. For each passenger in the test set, I have models like Logistic, DT, KNN Classifier, and XGBoost to predict whether or not they survived the sinking of the Titanic.